langchain-poc/openinference/requirements.txt
langchain
langchain-community
langchain-openai
opentelemetry-sdk
opentelemetry-api
opentelemetry-semantic-conventions
openai
python-dotenv
openinference-instrumentation-langchain
botocore
setuptools
